make ai endpoint locations consistent; update handler
